:recycle: [repositories] Move fixture `fetcher` to `tests.fixtures`
